wasm-gc: extract collect_* discovery walkers into types_discovery.rs
`types.rs` had grown to 2778 lines, half of which (~850 LoC) was
`collect_*` walkers — pure pre-discovery passes that mine type
strings + AST bodies for monomorphisations the registry needs to
allocate ahead of structural emit (`Vector<T>`, `Option<T>`,
`Result<T,E>`, `List<T>`, `Map<K,V>`, tuple shapes). Each family
has three flavours: `_from_str` (walks serialised types),
`_from_fn_body` + `_from_expr` (walk the typed AST).

Move them into a new sibling `types_discovery.rs`. `types.rs`
keeps the `TypeRegistry` impl, `aver_to_wasm`, and the type-name
normalisation utilities (`normalize_compound`,
`strip_inner_dotted_prefixes`, `rewrite_paren_tuples`,
`parse_map_kv`). Both modules import from each other (`types.rs`
imports the walkers; `types_discovery.rs` imports `TypeRegistry`
+ `normalize_compound`) — pure code-organisation refactor, zero
behaviour change.

  src/codegen/wasm_gc/
    types.rs              2778 → 1934  (-844)
    types_discovery.rs    new   867

All wasm-gc e2e fixtures + wasip2 unit / POC tests still green.
