+++ src/codegen/wasm_gc/mod.rs
@@ -20,6 +20,7 @@ mod module;
 #[cfg(test)]
 mod tests;
 mod types;
+mod types_discovery;
 mod wasip2_helpers;
 mod wasip2_imports;
 mod wat_helper;
